requirements.txt
PyYAML >= 6.0.0
requests >= 2.32.0
playwright >= 1.45.0
mysql-connector-python >= 9.0.0